# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Example APP",
        "version": "v1"
    },
    "host": "api-boilerplate.dev",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/x-www-form-urlencoded",
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/user/update": {
            "post": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Updates the current user",
                "description": "Updates the current user",
                "operationId": "update_current_user",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user/change_password": {
            "post": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Updates the current users password",
                "description": "Updates the current users password",
                "operationId": "change_current_user_password",
                "parameters": [
                    {
                        "name": "old_password",
                        "in": "formData",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "formData",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}